Separate true sentinels from sentinel-based common errors.
