+++ pkg/errorhelpers/custom_types.go
@@ -1,8 +1,6 @@
 package errorhelpers
 
 import (
-	"fmt"
-
 	"github.com/pkg/errors"
 )
 
@@ -21,51 +19,17 @@ var (
 	// ErrNotFound indicates that the requested object was not found.
 	ErrNotFound = errors.New("not found")
 
-	// ErrReferencedByAnotherObject indicates that the requested object cannot
-	// be removed because it is referred to / in use by another object.
-	ErrReferencedByAnotherObject = errors.New("referenced by another object")
-
 	// ErrInvariantViolation indicates that some internal invariant has been
 	// violated and the underlying component is in an inconsistent state.
 	ErrInvariantViolation = errors.New("invariant violation")
 
-	// ErrNoCredentials occurs if no credentials can be found.
-	ErrNoCredentials = errors.New("credentials not found")
+	// ErrReferencedByAnotherObject indicates that the requested object cannot
+	// be removed because it is referred to / in use by another object.
+	ErrReferencedByAnotherObject = errors.New("referenced by another object")
 
-	// ErrNoValidRole occurs if no valid role can be found for user.
-	ErrNoValidRole = errors.New("no valid role")
+	ErrNotAuthenticated = errors.New("not authenticated")
 
 	// ErrNotAuthorized occurs if credentials are found, but they are
 	// insufficiently authorized.
 	ErrNotAuthorized = errors.New("not authorized")
-
-	// ErrNoAuthzConfigured occurs if authorization is not implemented for a
-	// service. This is a programming error.
-	ErrNoAuthzConfigured = errors.New("service authorization is misconfigured")
 )
-
-// GenericNoValidRole wraps ErrNoValidRole with a generic error message
-func GenericNoValidRole() error {
-	return fmt.Errorf("Access for this user is not authorized: %w. Please contact a system administrator.",
-		ErrNoValidRole)
-}
-
-// NewErrNotAuthorized wraps ErrNotAuthorized into an explanation.
-func NewErrNotAuthorized(explanation string) error {
-	return fmt.Errorf("%w: %s", ErrNotAuthorized, explanation)
-}
-
-// NewErrNoCredentials wraps ErrNoCredentials into an explanation.
-func NewErrNoCredentials(explanation string) error {
-	return fmt.Errorf("%w: %s", ErrNoCredentials, explanation)
-}
-
-// NewErrInvariantViolation wraps ErrInvariantViolation into an explanation.
-func NewErrInvariantViolation(explanation string) error {
-	return fmt.Errorf("%w: %s", ErrInvariantViolation, explanation)
-}
-
-// NewErrInvalidArgs wraps ErrInvalidArgs into an explanation.
-func NewErrInvalidArgs(explanation string) error {
-	return fmt.Errorf("%w: %s", ErrInvalidArgs, explanation)
-}
